feat: add tokmd-types crate and update dependencies across the project
