+++ crates/tokmd-model/Cargo.toml
@@ -8,3 +8,4 @@ anyhow = "1.0.100"
 serde = { version = "1.0.228", features = ["derive"] }
 tokei = { version = "14.0.0", default-features = false }
 tokmd-config = { path = "../tokmd-config" }
+tokmd-types = { path = "../tokmd-types" }
