+++ crates/tokmd-types/Cargo.toml
@@ -1,13 +1,8 @@
 [package]
-name = "tokmd-format"
+name = "tokmd-types"
 version = "1.0.0"
 edition = "2024"
 
 [dependencies]
-anyhow = "1.0.100"
-blake3 = "1.8.3"
-csv = "1.4.0"
 serde = { version = "1.0.228", features = ["derive"] }
-serde_json = "1.0.149"
 tokmd-config = { path = "../tokmd-config" }
-tokmd-types = { path = "../tokmd-types" }
